Use flask server instead of tornado
