+++ qcfractal/requirements.txt
@@ -1,6 +1,6 @@
 blinker==1.4
 Click==7.0
-Flask==1.0.2
+Flask==1.0.4
 Flask-JWT-Extended==3.24.1
 Flask-Mail==0.9.1
 flask-marshmallow==0.9.0
